requirements.txt
django
requests
djangorestframework
dj-rest-auth[with_social]
whitenoise
Pillow
aiohttp
django-cors-headers
django-filter
openpyxl
django-extensions
